Allow for all(varnames %in% cd.names) to be ok
When more rasters where provided to maxlike then are used in the formula the prediction function should still work. This was not the case anymore as it was checked if all provided rasters/newdata where present in the formula. By adding a second clause checking for the inverse case I believe that the problem with interaction terms Roeland Kindt experienced and the problem reported to me by David Sabadin are both handled.